Moving 'tree' to kernel.
